- [WIP] move transaction logic to lib/invite, almost complete flow
