+++ src/bridge/views/InviteTransactions.js
@@ -68,245 +68,6 @@ class InviteTransactions extends React.Component {
     this.setState({progress: msg+'...'});
   }
 
-  askForFunding(address, amount, current) {
-    this.updateProgress(`Please make sure ${address} has at least ${amount} wei, we'll continue once that's true. Current balance: ${current}. Waiting`);
-  }
-
-  async startTransactions() {
-    const { point } = this.state;
-
-    if (Nothing.hasInstance(this.props.web3)) {
-      throw BRIDGE_ERROR.MISSING_WEB3;
-    }
-    const web3 = this.props.web3.value;
-
-    if (Nothing.hasInstance(this.props.contracts)) {
-      throw BRIDGE_ERROR.MISSING_CONTRACTS;
-    }
-    const contracts = this.props.contracts.value;
-
-    if (Nothing.hasInstance(this.props.wallet)) {
-      throw BRIDGE_ERROR.MISSING_WALLET;
-    }
-    const inviteWallet = this.props.wallet.value;
-    const inviteAddress = addressFromSecp256k1Public(inviteWallet.publicKey);
-    console.log('working as', inviteAddress);
-
-    let newUrbitWallet = await urbitWalletFromTicket(this.state.ticket, point);
-
-    // transfer from invite wallet to new wallet
-
-    let transferTx = azimuth.ecliptic.transferPoint(contracts, point, newUrbitWallet.ownership.keys.address);
-    transferTx.gas = 500000; //TODO can maybe be lower?
-
-    // ping gas tank with txs if needed
-
-    let inviteNonce = await web3.eth.getTransactionCount(inviteAddress);
-    transferTx.nonce = inviteNonce++;
-    transferTx.gasPrice = 20000000000; //NOTE we pay the premium for faster ux
-    transferTx.from = inviteAddress;
-
-    //NOTE using web3.eth.accounts.signTransaction is broken (1.0.0-beta51)
-    let transferStx = new Tx(transferTx);
-    transferStx.sign(inviteWallet.privateKey);
-    let rawTransferStx = '0x'+transferStx.serialize().toString('hex');
-
-    const transferCost = transferTx.gas * transferTx.gasPrice;
-    await this.ensureFundsFor(web3, point, inviteAddress, transferCost, [rawTransferStx]);
-
-    // send transaction
-    this.updateProgress(PROGRESS_STATES.CLAIMING);
-    await this.sendAndAwaitConfirm(web3, [rawTransferStx]);
-
-    //
-    // we're gonna be operating as the new wallet from here on out, so change
-    // the relevant values
-    //
-    this.props.setUrbitWallet(Just(newUrbitWallet));
-    const newAddress = newUrbitWallet.ownership.keys.address;
-
-    // configure management proxy
-
-    let managementTx = azimuth.ecliptic.setManagementProxy(contracts, point, newUrbitWallet.management.keys.address);
-    managementTx.gas = 200000;
-    managementTx.nonce = 0;
-
-    // configure networking public keys
-    //TODO feels like more of this logic should live in a lib?
-    let networkSeed = await attemptSeedDerivation(true, {
-      walletType: WALLET_NAMES.TICKET,
-      urbitWallet: Just(newUrbitWallet),
-      pointCursor: Just(point),
-      pointCache: {[point]: {keyRevisionNumber: 0}}
-    });
-    if (Nothing.hasInstance(networkSeed)) {
-      throw new Error('wtf network seed not derived');
-    }
-    networkSeed = networkSeed.value;
-    const networkKeys = kg.deriveNetworkKeys(networkSeed);
-
-    let keysTx = azimuth.ecliptic.configureKeys(
-      contracts,
-      point,
-      addHexPrefix(networkKeys.crypt.public),
-      addHexPrefix(networkKeys.auth.public),
-      1,
-      false
-    );
-    keysTx.gas = 150000;
-    keysTx.nonce = 1;
-
-    //TODO  refactor this process into function
-    let totalCost = 0;
-    let txs = [managementTx, keysTx];
-    txs = txs.map(tx => {
-      tx.gasPrice = 20000000000;
-      tx.from = newAddress;
-      totalCost = totalCost + (tx.gasPrice * tx.gas);
-      return tx;
-    });
-
-    let rawStxs = txs.map(tx => {
-      let stx = new Tx(tx);
-      stx.sign(Buffer.from(newUrbitWallet.ownership.keys.private, 'hex'));
-      return '0x'+stx.serialize().toString('hex');
-    });
-
-    await this.ensureFundsFor(web3, point, newAddress, totalCost, rawStxs);
-
-    this.updateProgress(PROGRESS_STATES.CONFIGURING);
-    await this.sendAndAwaitConfirm(web3, rawStxs);
-
-    // if non-trivial eth left in invite wallet, transfer to new ownership
-    this.updateProgress(PROGRESS_STATES.CLEANING);
-    let balance = await web3.eth.getBalance(inviteAddress);
-    const gasPrice = 20000000000;
-    const gasLimit = 21000;
-    const sendEthCost = gasPrice * gasLimit;
-    if (balance > sendEthCost) {
-      const value = (balance - sendEthCost);
-      console.log('sending', value);
-      const tx = {
-        to: newAddress,
-        value: value,
-        gasPrice: gasPrice,
-        gas: gasLimit,
-        nonce: inviteNonce++
-      }
-      let stx = new Tx(tx);
-      stx.sign(inviteWallet.privateKey);
-      const rawTx = '0x'+stx.serialize().toString('hex');
-      web3.eth.sendSignedTransaction(rawTx).catch(err => {
-        console.log('error sending value tx, who cares', err);
-      });
-      console.log('sent old balance');
-    }
-
-    // proceed without waiting for confirm
-    this.updateProgress(PROGRESS_STATES.DONE);
-    this.props.setUrbitWallet(Just(newUrbitWallet));
-    //TODO forward to "all done!" screen
-  }
-
-  async ensureFundsFor(web3, point, address, cost, signedTxs) {
-    this.updateProgress(PROGRESS_STATES.FUNDING);
-    let balance = await web3.eth.getBalance(address);
-
-    if (cost > balance) {
-
-      try {
-
-        const fundsRemaining = await tank.remainingTransactions(point);
-        if (fundsRemaining < signedTxs.length) {
-          throw new Error('request invalid');
-        }
-
-        const res = await tank.fundTransactions(signedTxs);
-        if (!res.success) {
-          throw new Error('request rejected');
-        } else {
-          await waitForTransactionConfirm(web3, res.txHash);
-          let newBalance = await web3.eth.getBalance(address);
-          console.log('funds have confirmed', balance >= cost, balance, newBalance);
-        }
-
-      } catch (e) {
-
-        console.log('funding failed', e);
-        await this.waitForBalance(web3, address, cost);
-
-      }
-
-    } else {
-      console.log('already have sufficient funds');
-    }
-  }
-
-  // resolves when address has at least minBalance
-  //
-  async waitForBalance(web3, address, minBalance) {
-    console.log('awaiting balance', address, minBalance);
-    return new Promise((resolve, reject) => {
-      let oldBalance = null;
-      const checkForBalance = async () => {
-        const balance = await web3.eth.getBalance(address);
-        if (balance >= minBalance) {
-          resolve();
-        } else {
-          if (balance !== oldBalance) {
-            this.askForFunding(address, minBalance, balance);
-          }
-          setTimeout(checkForBalance, 13000);
-        }
-      };
-      checkForBalance();
-    });
-  }
-
-  async sendAndAwaitConfirm(web3, signedTxs) {
-    await Promise.all(signedTxs.map(tx => {
-      console.log('sending...');
-      return new Promise((resolve, reject) => {
-        web3.eth.sendSignedTransaction(tx).then(res => {
-          console.log('sent, now waiting for confirm!', res.transactionHash);
-          waitForTransactionConfirm(web3, res.transactionHash)
-          .then(resolve);
-        }).catch(async err => {
-          // if there's an error, check if it's because the transaction was
-          // already confirmed prior to sending.
-          // this is really only the case in local dev environments.
-          const txHash = web3.utils.keccak256(tx);
-          console.log(err.message);
-          if (err.message.slice(0,54) ===
-              "Returned error: the tx doesn't have the correct nonce.") {
-            console.log('nonce error, awaiting confirm', err.message.slice(55));
-            //TODO max wait time before assume failure?
-            let res = await waitForTransactionConfirm(web3, txHash);
-            if (res) resolve();
-            else reject(new Error('Unexpected tx failure'));
-          } else {
-            const confirmed = await isTransactionConfirmed(web3, txHash);
-            console.log('error, but maybe confirmed:', confirmed);
-            if (confirmed) resolve();
-            else reject(err);
-          }
-        });
-      });
-    }));
-  }
-
-  awaitForResult(desired, retryTime, func) {
-    return new Promise((resolve, reject) => {
-      const retry = async () => {
-        const result = await func();
-        console.log('result vs desired', result, desired);
-        if (result === desired) resolve();
-        else setTimeout(retry, retryTime);
-      };
-      retry();
-    });
-  }
-
   render() {
 
     const { point } = this.state;
